requirements.txt
streamlit
replicate
transformers
pytesseract
pillow
pdfplumber